+++ pkgs/artisanal/lib/style.dart
@@ -111,7 +111,7 @@ export 'src/style/list.dart'
 export 'src/layout/layout.dart' show Layout, WhitespaceOptions;
 export 'src/style/ranges.dart'
     show StyleRange, styleRanges, Ranges, cutAnsiByCells;
-export 'src/style/blending.dart' show blend1D, blend2D;
+export 'src/style/blending.dart' show blend1D, blend2D, blendColor;
 export 'src/style/writer.dart'
     show
         Writer,